Include ThreadMan, move threading-related stuff into there
