+++ Managers/meson.build
@@ -20,7 +20,7 @@ sources += files(
 'ThreadMan.cpp',
 'TimerMan.cpp',
 'UInputMan.cpp',
-'WindowMan.cpp'
+'ThreadMan.cpp',
 )
 luaman_cpp_args = preprocessor_flags
 defaults_override = []
